% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{list_crs}
\alias{list_crs}
\title{Print the currently accepted EPSG codes.}
\usage{
list_crs(...)
}
\arguments{
\item{...}{Not currently used.}
}
\value{
(Invisible) Vector of character strings.
}
\description{
Convenience function primarily used internally by \code{osdatahub}.
}
\examples{
list_crs()

}
