% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdatahub-package.R
\docType{package}
\name{osdatahub-package}
\alias{osdatahub}
\alias{osdatahub-package}
\title{\code{osdatahub}: Easier Interaction with the Ordnance Survey Data Hub}
\description{
Ordnance Survey ('OS') is the national mapping agency for Great Britain and
produces a large variety of mapping and geospatial products. Much of OS's
data is available via the OS Data Hub \url{https://osdatahub.os.uk/}, a platform
that hosts both free and premium data products. \code{osdatahub} provides a
user-friendly way to access, query, and download these data in \code{R}.
}
\author{
\strong{Maintainer}: Chris Jochem \email{chris.jochem@os.uk}

Other contributors:
\itemize{
  \item Ordnance Survey Ltd [copyright holder, funder]
}

}
\keyword{internal}
